requirements.txt
Django==2.2.2
djangorestframework==3.9.4
gevent==1.4.0
greenlet==0.4.15
kafka-python==1.4.6
python-dateutil==2.8.0
pytz==2019.1
redis==3.2.1
six==1.12.0
sqlparse==0.3.0
websocket==0.2.1
websocket-client==0.56.0
